fix: compose IntegrationSlice into unified store (#534)
The integrations slice lived in src/store/ as a standalone Zustand
instance, separate from the main composed store in src/lib/store/store.ts.
This caused integration state to be invisible to any selector on useStore,
with no shared persistence, no cross-slice access, and duplicated
notification typing.

Changes:
- Create src/lib/store/integrations/ following the existing slice pattern
  (types, slice, selectors, api, tests) matching other slices like health,
  collaboration, and financing.
- Rename health-related fields to integrationHealth* prefix to avoid
  naming collision with HealthSlice's healthLoading (object vs boolean).
- Compose createIntegrationSlice into the main persist(...) store and add
  IntegrationSlice to the StoreState intersection type.
- Update IntegrationsOverview to use useStore instead of useIntegrationStore.
- Update ConnectionForm and ConnectionCard to import types from new location.
- Add backward-compatible re-exports in src/store/ for existing imports.
- Add regression test verifying integration + auth state are reachable
  from the same store instance.

Verification:
- TypeScript typecheck passes (npx tsc --noEmit)
- All 194 tests pass including 7 new integration slice tests
