+++ project-portal/project-portal-web/src/lib/store/store.ts
@@ -22,6 +22,8 @@ import type { GeospatialSlice } from "./geospatial/geospatial.types";
 import { createGeospatialSlice } from "./geospatial/geospatialSlice";
 import type { ReportsSlice } from "./reports/reports.slice";
 import { createReportsSlice } from "./reports/reports.slice";
+import type { IntegrationSlice } from "./integrations/integration.types";
+import { createIntegrationSlice } from "./integrations/integrationSlice";
 
 // Unified store state type
 export type StoreState = AuthSlice &
@@ -32,7 +34,8 @@ export type StoreState = AuthSlice &
   NotificationsSlice &
   FinancingSlice &
   GeospatialSlice &
-  ReportsSlice;
+  ReportsSlice &
+  IntegrationSlice;
 
 // Helper to check if token is expired or about to expire (60s buffer)
 const isTokenExpiringSoon = (expiresIn: number | null): boolean => {
@@ -53,6 +56,7 @@ export const useStore = create<StoreState>()(
       ...createFinancingSlice(...args),
       ...createGeospatialSlice(...args),
       ...createReportsSlice(...args),
+      ...createIntegrationSlice(...args),
     }),
     {
       name: "project-portal-store",
