+++ project-portal/project-portal-web/src/store/store.ts
@@ -1,12 +1,12 @@
 /**
- * Unified store entry. Integrations and notifications slices are registered here.
- * Extend with auth, projects, collaboration, search as needed.
+ * @deprecated Import from '@/lib/store/store' instead.
+ * This file is kept for backward compatibility only.
+ * Integration state is now composed into the unified useStore hook.
  */
-
-export { useIntegrationStore } from "./integrationSlice";
-export type { IntegrationSlice } from "./integrationSlice";
-export * from "./integration.selectors";
-export * from "./integration.types";
+export { useStore } from '@/lib/store/store';
+export type { IntegrationSlice } from '@/lib/store/integrations/integration.types';
+export * from '@/lib/store/integrations/integration.selectors';
+export * from '@/lib/store/integrations/integration.types';
 
 export type {
   Notification,
